Add Packet class, with payload-buffer and size...
Add `Packet` class, with the following attributes:

 - `command_`
 - `payload_buffer_`
 - `payload_length_`

__NB__ During the `PacketParser::parse` method, int the case of a
successful packet parse operation, the `packet.payload_buffer_` is set
to the location in `buffer` where the payload is found during parsing,
while `packet.payload_length_` is set to the length of the payload.
This means that the pointer `packet.payload_buffer_` is only valid as
long as the `buffer` _(passed to the `PacketParser::parse`)_ method is
valid.

If the packet parse is unsuccessful, return `false`.  In the case of an
unsuccessful parse attempt, the state of attributes the attributes of
`packet` are _undefined_.