Support request _and_ response packet types...
Support request _and_ response packet types.

In other words, support commands values in the range `0-127`, stored in
a byte _(i.e., 8 bits)_.  The _Most-Significant-Bit (MSB)_ is used to
specify the type of the command as either a _request (`1`)_ or _response
(`0`)_.

Changes to packet grammar
=========================

The packet grammar to allow values between `0-255` for the
command byte, in contrast to before this commit, where only values in
the range `128-255` were permitted _(i.e., only requests were allowed)_.

Changes to packet
=================

The `Packet` class now contains a `type_` field which is set through the
`command(command_byte_with_type_msb)` setter method, which also strips the MSB
before setting the `command` field.