[WIP] Add initial PacketSocket draft
Add initial draft of `PacketSocket` class, which processes events
related to sending, receiving, and validating packets on an interface.

Add `dump_events_header.py` script to extract the labels of events
defined in a Ragel state-chart.  The script can be used to generate the
`PacketSocketEvents.h` header based on the `packet_socket_fsm.rl` file
to provide a reverse mapping from each event character to the
label for the corresponding event.