Add PacketAllocator, restructure build layout
Add `PacketAllocator` to allocate packets, provided an `Allocator` type,
where the `Allocator` implements the following interface:

 - `uint_8 *alloc()`: Allocate a buffer.
 - `void free(uint_8 *)`: Free a buffer that was allocated by `alloc()`.

__NB__ Currently, only fixed size buffers of a pre-defined size may be
created, as there is no argument to the `alloc` function to specify the
size of the buffer.  The default `Allocator` type for the
`PacketAllocator` is the `FixedSizeBufferPool`, which allocates buffers
of a fixed size, where the size is defined at compile-time.

Restructure build layout to build all binaries in a `build` directory in
the root of the repository.

Notes
=====

TODO - We could probably add a size argument to the `Allocator::alloc`
method.  In the case of the `FixedSizeBufferPool`, we would just return
a `NULL` pointer if the requested buffer size was too large.  Otherwise,
we would just return a fixed-size buffer, which would be at least as
large as the requested size.  This would lead to wasted memory in some
cases, but should be sufficient for our use case and will keep the
interface general enough to support allocators applicable to
environments that are not memory-constrained.