Enable socket use with a buffer pool size of 1...
Enable socket use with a payload buffer pool size of one.  To accomplish
this, we only allocate new payload buffer for the parser to use when the
next byte is available on the input stream.  Prior to this commit, the
parser held a packet payload buffer, which meant that the parser always
tied up a buffer, even when no new data was ready to be parsed.

As of this commit, the packet parser passes ownership of the payload
buffer from a successfully parsed data packet to the receiving packet of
the socket. If the parser's packet has no buffer allocated and there is
no buffer left in the allocation pool, no data will be read from the
input stream. However, as long as data is available on the input stream,
buffer allocation will be retried until a buffer becomes available in
the allocation pool.  This allows parsing to resume once a buffer has
been released at some point down the processing line.