fix(appveyor): fix Miniconda path
As of `nadamq==0.18`, the paths `C:\Miniconda3` and `C:\Miniconda3-x64`
corresponded to the path to the latest Miniconda Python 3 root environment
for 32-bit Windows and 64-bit Windows, respectively.  However, as of this
commit, these paths no longer exist on Appveyor.

In this commit, explicitly use the Python 3.6 Miniconda environment paths,
i.e., `C:\Miniconda36` and `C:\Miniconda36-x64`.